+++ libs/schematics/src/generators/library/helpers/index.ts
@@ -1,1 +1,5 @@
+export * from './clean-angular-library.helper';
+export * from './clean-nest-library.helper';
+export * from './create-secondary-entrypoints.helper';
+export * from './create-src-index.helper';
 export * from './normalized-schema.helper';